Set poisson Lambda as a float

+++ appli/education/distributed_algorithm/poisson_clock.h
@@ -1,7 +1,7 @@
 #ifndef POISSON_CLOCK_H
 #define POISSON_CLOCK_H
 
-unsigned int poisson_delay_in_ticks(unsigned int lambda);
+unsigned int poisson_delay_in_ticks(double lambda);
 
 int poisson_delay(int argc, char **argv);
